Changes chunkpool to maintain reference counts.
Rather than simply acquiring a or releasing a chunk,
the code now increments and decrements a reference count:
	 - chunkpool_acquire() sets refcnt to 1.
	 - any function that stores a reference to the chunk after
	 the function is out of scope (i.e. copies it to a static)
	 MUST increment the refcnt (unless it is the function that called
	 chunkpool_acquire())
	 - any time the refcnt is incremented (including chunkpool_acquire),
	 there must be a corresponding call to chunkpool_decref() after the
	 reference is on longer used.
	 - Example: function A() acquires a chunk and passes it to function
	   B(chunk). A() does not save any references to the chunk, but
	   before B() returns, B() copies the pointer to some static
	   location. B() must call chunkpool_incref(chunk). Since A()
	   created the chunk, A() must call chunkpool_decref() when it
	   returns.

The main use of this will be to facilitate creating a single packet to
transmit to multiple destinations. Each time a transmission finishes,
the refcnt is decremented; when they all finish, the refcnt is zero and
the chunk is considered free.