mpc/serialcomm/comm: Add serialcomm receiver
Adds serial comm receiver with a [probably really stupid] frame
synchronizer.  The synchronizer runs constantly, and begins trying to
syncrhonize whenever it sees 0xFF.  If a frame is being received while
synchronization is running,  the data is received as normally. When the
syncrhonizer detects a valid 4-byte 'start sequence' (0xff, dest
address, pkt length, check), any active RX is ended, and a new RX is
started.  Since the syncrhonizer restarts whenever 0xFF is received,
none of the bytes in the frame header can be 0xFF - this means the size
is limited to 0xFE and the check must not be 0xFF.  The check is defined
as: ((destination address)^(size))0xFE; it is a simple xor of the
address and size bytes with the 1 bit cleared.