sercom: Unify interrupt handling; works on SAML21 and SAMD51
(not that I have tested on SAML21, or with interrupt line 3...)

SAML21 has one interrupt line per sercom; SAMD51 has four. The first
four interrupt flags use dedicated lines while the remainder are
multiplexed in the 4th. SAML21 multiplexes all of them in one handler.

This moves all of the handling of those to the sercom driver.